requirements.txt
telethon>=1.28.5
cryptg>=0.4.0
psutil>=5.9.5
python-dotenv>=1.0.0
speedtest-cli>=2.1.3
